requirements.txt
asgiref==3.5.0
attrs==21.4.0
bootstrap4==0.1.0
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
Currency==0.1.2
distlib==0.3.6
Django==4.0.2
django-admin-thumbnails==0.2.6
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-humanize==0.1.2
django-jet-reboot==1.3.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-paypal==1.1.2
django-taggit==3.0.0
filelock==3.8.0
humanize==4.0.0
idna==3.3
isodate==0.6.1
lxml==4.8.0
Pillow==9.0.0
platformdirs==2.5.1
pytz==2021.3
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
virtualenv==20.16.5
zeep==4.1.0